backend/requirements.txt
# FastAPI and ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Core Dependencies
python-dotenv==1.0.0
google-generativeai==0.3.1
pillow==10.1.0
pymongo==4.6.1
bcrypt==4.0.1
PyJWT==2.8.0
pdf2image==1.16.3
tenacity==8.2.3
tqdm==4.66.1
requests==2.31.0

# Production Dependencies
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2  # Async MongoDB driver
redis==5.0.1  # Caching
celery==5.3.4  # Background tasks
prometheus-client==0.19.0  # Metrics
structlog==23.2.0  # Structured logging

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # Testing client 